@@ -5,69 +5,82 @@ description = "Retro Battleship game with FastAPI, HTMX, and modern DevOps."
 authors = ["Your Name <you@example.com>"]
 readme = "README.md"
 packages = [{ include = "src" }]
-# NOTE: remove `requires-python`; Poetry uses `python` under dependencies.
 
 [tool.poetry.dependencies]
-python = "^3.12"
+# Support 3.12 and 3.13, but you're currently on 3.13
+python = ">=3.12,<3.14"
+
+
+
 # Web framework
 fastapi = ">=0.111,<1.0"
 uvicorn = { extras = ["standard"], version = ">=0.30,<1.0" }
 httpx = ">=0.27,<0.28"
 jinja2 = ">=3.1,<4.0"
 python-multipart = ">=0.0.12,<0.1"
-# Database
+
+# Database (async via psycopg3) — asyncpg removed
 psycopg = { version = ">=3.2,<4.0", extras = ["binary"] }
 sqlalchemy = ">=2.0,<3.0"
 alembic = ">=1.13,<2.0"
-asyncpg = ">=0.29,<0.30"
-redis = ">=5.0,<6.0"
+redis = { version = ">=5.0,<6.0", optional = true }
+
 # Security & Auth
 passlib = { extras = ["argon2"], version = ">=1.7,<2.0" }
 email-validator = ">=2.1,<3.0"
 python-jose = { extras = ["cryptography"], version = ">=3.3,<4.0" }
 python-decouple = ">=3.8,<4.0"
-fastapi-users = { extras = ["sqlalchemy", "oauth"], version = ">=13.0,<14.0" }
-# Core pydantic (was in requirements but missing here)
+fastapi-users = { extras = ["sqlalchemy", "oauth"], version = ">=13.0,<15.0" }
 pydantic = ">=2.6,<3.0"
 ecdsa = ">=0.19.2"
+
+[tool.poetry.extras]
+redis = ["redis"]
 
 [tool.poetry.group.dev.dependencies]
 # Code Quality
 black = "25.1.0"
 ruff = "0.5.6"
 mypy = "1.17.0"
+testcontainers = { version = "^4.9.0", extras = ["postgres"] }
+PyJWT = "^2.9.0"
+types-PyJWT = "*"
+
 # Testing
 pytest = "8.4.1"
 pytest-asyncio = "1.1.0"
 pytest-cov = "5.0.0"
 coverage = { extras = ["toml"], version = "7.6.1" }
+
 # DB Testing
 pytest-postgresql = "6.0.0"
 factory-boy = "3.3.0"
 faker = "22.0.0"
+
 # Dev Tools
 pre-commit = "4.0.1"
 pgcli = "4.0.1"
+
 # Performance
 locust = "2.37.14"
-# Type stubs
+
+# Type stubs (keep only the ones you actually need)
 types-requests = "*"
 types-python-dateutil = "*"
 types-redis = "*"
-asyncpg-stubs = "*"
-sqlalchemy-stubs = "*"
 
 [tool.black]
 line-length = 88
-target-version = ['py312']
+target-version = ['py313']
 
 [tool.ruff]
 line-length = 88
-target-version = "py312"
+target-version = "py313"
 
 [tool.ruff.lint.per-file-ignores]
-"tests/**/*.py" = ["S101"]
-
+"tests/**/*.py" = ["S101", "PT001"]
+"tests/conftest.py" = ["PLC0415"]
+"tests/db_bootstrap_test.py" = ["PLC0415"]
 
 [tool.ruff.lint]
 select = ["ALL"]
@@ -75,7 +88,7 @@ ignore = ["D203", "D212", "E501"]
 fixable = ["ALL"]
 
 [tool.mypy]
-python_version = "3.12"
+python_version = "3.13"
 strict = true
 warn_unused_configs = true
 disallow_untyped_defs = true
@@ -87,6 +100,10 @@ disable_error_code = ["misc"]
 module = ["decouple"]
 ignore_missing_imports = true
 
+[[tool.mypy.overrides]]
+module = ["pytest", "pytest.*", "testcontainers", "testcontainers.*", "jwt"]
+ignore_missing_imports = true
+
 [tool.isort]
 profile = "black"
 line_length = 88
@@ -96,17 +113,16 @@ known_third_party = [
     "httpx",
     "jinja2",
     "redis",
-    "asyncpg",
     "sqlalchemy",
     "alembic",
 ]
 
 [tool.pylint.MESSAGES_CONTROL]
 disable = [
-    "redefined-outer-name",  # Common in pytest fixtures
+    "redefined-outer-name",
 ]
 
 [tool.pylint.'tests.*']
 disable = [
-    "redefined-outer-name",  # Only disable in test files
+    "redefined-outer-name",
 ]
